# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "-",
        "title": "Documenting Whitebook API",
        "contact": {},
        "version": "1"
    },
    "basePath": "/",
    "paths": {
        "/advertisement/create": {
            "post": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/advertisement/delete/{advertisementsId}": {
            "delete": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/advertisement/update": {
            "patch": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/advertisement/{advertisementsId}": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/advertisements": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/book-preview-image/create": {
            "post": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/book-preview-image/delete/{bookPreviewImageId}": {
            "delete": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/book-preview-image/update": {
            "patch": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/book-preview-image/{bookPreviewImageId}": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/book-preview-images": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "responses.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "bearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}